fix(site): rename /docs/guides/workflows route to /docs/guides/standalone-tasks
The guide page was retitled to "Standalone Tasks" but the URL slug still
used the legacy "workflows" path. This renames the route directory, updates
all nav/sitemap/cross-link references, and adds a client-side redirect page
at the old path so existing external links don't 404.

- Rename workflows/ directory to standalone-tasks/
- Update docs.ts nav href
- Update sitemap.ts entry
- Update cross-links in connections and integrations guide pages
- Add redirect page at old /docs/guides/workflows path
- Add vitest + tests for docs nav and sitemap entries
