+++ apps/site/src/content/docs.ts
@@ -28,7 +28,7 @@ export const docsNav: DocSection[] = [
     title: "Guides",
     pages: [
       { title: "Creating Tasks", href: "/docs/guides/creating-tasks" },
-      { title: "Standalone Tasks", href: "/docs/guides/workflows" },
+      { title: "Standalone Tasks", href: "/docs/guides/standalone-tasks" },
       { title: "Scheduled Tasks", href: "/docs/guides/scheduled-tasks" },
       { title: "Connecting Repos", href: "/docs/guides/connecting-repos" },
       { title: "Review Agents", href: "/docs/guides/review-agents" },
